class structure redesigned
