+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sportsportal",
-  "version": "0.5.99-SNAPSHOT",
+  "version": "0.6.0",
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
