+++ client/src/boot/auth/Confirmation.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MainFrame from "../frame/MainFrame";
+import MainFrame from '../frame/MainFrame';
 import apiUrl from '../constants'
 import axios from 'axios';
 import qs from 'qs';
