+++ client/src/boot/frame/MainFrame.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import ModalFade from '../../util/components/ModalFade'
+import ModalFade from '../../util/components/ModalFade';
 import MainContainer from './sections/MainContainer';
-import Login from './modal/Login'
-import Message from './modal/Message'
+import Login from './modal/Login';
+import Message from './modal/Message';
 import Registration from './modal/Registration'
 import Header from './sections/Header';
 import Footer from './sections/Footer';
