+++ client/src/boot/frame/sections/MainContainer.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 import './MainContainer.css';
 
 export default function MainContainer(props) {
