+++ client/src/main/playground/PlaygroundInfo.css
@@ -1,0 +1,29 @@
+.container.content-container {
+    padding: 0;
+}
+
+.container.content-container .row.info-row {
+    padding: 10px 0;
+    margin-top: 15px;
+    margin-bottom: 15px;
+    border-radius: .25rem;
+    background-color: #f8f9fa;
+}
+
+.container.content-container .row.info-row .row-h {
+    font-weight: 300;
+}
+
+.container.content-container .row.header-info-row h1.header-h {
+    margin-bottom: 0;
+}
+
+.container.content-container .row.feature-info-row .info-h.info-infra {
+    padding: .25rem .75rem;
+}
+
+.container.content-container .row.feature-info-row ul li {
+    padding: .5rem .75rem;
+    background: inherit;
+    font-weight: 300;
+}
