+++ client/src/util/components/CustomCheckbox.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 /**
  * CustomCheckbox by bootstrap custom-control.
