+++ client/src/util/components/PhotoCarousel.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 /**
  * PhotoCarousel by bootstrap carousel.
