+++ client/src/util/components/StarRate.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 /**
  * StarRate by font awesome carousel.
